Connect drill down UI with BE insights gql handler
