+++ client/web/src/insights/components/insights-view-grid/components/backend-insight/components/drill-down-filters-action/DrillDownFiltersPanel.module.scss
@@ -1,5 +1,3 @@
-@import '~@reach/menu-button/styles.css';
-
 .filter-button {
     position: relative;
