+++ client/web/src/insights/core/backend/api/get-backend-insight-by-id.ts
@@ -2,11 +2,11 @@ import { Observable, of, throwError } from 'rxjs'
 import { map, switchMap } from 'rxjs/operators'
 
 import { fetchBackendInsights } from '../requests/fetch-backend-insights'
-import { BackendInsightData } from '../types'
+import { BackendInsightData, BackendInsightFilters } from '../types'
 import { createViewContent } from '../utils/create-view-content'
 
-export function getBackendInsightById(id: string): Observable<BackendInsightData> {
-    return fetchBackendInsights([id]).pipe(
+export function getBackendInsightById(id: string, filters?: BackendInsightFilters): Observable<BackendInsightData> {
+    return fetchBackendInsights([id], filters).pipe(
         switchMap(backendInsights => {
             if (backendInsights.length === 0) {
                 return throwError(new Error("We couldn't find insight"))
