+++ client/web/src/insights/core/backend/requests/fetch-backend-insights.ts
@@ -11,6 +11,7 @@ import {
     SubjectSettingsResult,
     SubjectSettingsVariables,
 } from '../../../../graphql-operations'
+import { BackendInsightFilters } from '../types'
 
 const insightFieldsFragment = gql`
     fragment InsightFields on Insight {
@@ -19,17 +20,21 @@ const insightFieldsFragment = gql`
         description
         series {
             label
-            points {
+            points(excludeRepoRegex: $excludeRepoRegex, includeRepoRegex: $includeRepoRegex) {
                 dateTime
                 value
             }
         }
     }
 `
-export function fetchBackendInsights(insightsIds: string[]): Observable<InsightFields[]> {
+
+export function fetchBackendInsights(
+    insightsIds: string[],
+    filters?: BackendInsightFilters
+): Observable<InsightFields[]> {
     return requestGraphQL<InsightsResult>(
         gql`
-            query Insights($ids: [ID!]!) {
+            query Insights($ids: [ID!]!, $includeRepoRegex: String, $excludeRepoRegex: String) {
                 insights(ids: $ids) {
                     nodes {
                         ...InsightFields
@@ -38,7 +43,11 @@ export function fetchBackendInsights(insightsIds: string[]): Observable<InsightF
             }
             ${insightFieldsFragment}
         `,
-        { ids: insightsIds }
+        {
+            ids: insightsIds,
+            excludeRepoRegex: filters?.excludeRepoRegexp ?? null,
+            includeRepoRegex: filters?.includeRepoRegexp ?? null,
+        }
     ).pipe(
         map(dataOrThrowErrors),
         map(data => data.insights?.nodes ?? []),
