+++ client/web/src/insights/hooks/use-parallel-requests/use-parallel-request.ts
@@ -101,6 +101,10 @@ export function createUseParallelRequestsHook<T>({ maxRequests } = { maxRequests
         useEffect(() => {
             const cancelStream = new Subject<boolean>()
 
+            console.log('Start fetch use parallel')
+
+            setState({ data: undefined, loading: true, error: undefined })
+
             const event: Request<D> = {
                 request,
                 // Makes cancel stream a hot observable
